# agent/internal/api/generated/api.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
)

// Defines values for DockerActionRequestAction.
const (
	Restart DockerActionRequestAction = "restart"
	Start   DockerActionRequestAction = "start"
	Stop    DockerActionRequestAction = "stop"
)

// Defines values for HealthStatusStatus.
const (
	Degraded HealthStatusStatus = "degraded"
	Error    HealthStatusStatus = "error"
	Ok       HealthStatusStatus = "ok"
)

// Defines values for PowerActionRequestAction.
const (
	Reboot   PowerActionRequestAction = "reboot"
	Shutdown PowerActionRequestAction = "shutdown"
	Sleep    PowerActionRequestAction = "sleep"
)

// DockerActionRequest defines model for DockerActionRequest.
type DockerActionRequest struct {
	Action DockerActionRequestAction `json:"action"`
}

// DockerActionRequestAction defines model for DockerActionRequest.Action.
type DockerActionRequestAction string

// DockerActionResponse defines model for DockerActionResponse.
type DockerActionResponse struct {
	Message string `json:"message"`
	Success bool   `json:"success"`
}

// DockerContainer defines model for DockerContainer.
type DockerContainer struct {
	Id     string `json:"id"`
	Image  string `json:"image"`
	Name   string `json:"name"`
	Ports  string `json:"ports"`
	Status string `json:"status"`
	Uptime string `json:"uptime"`
}

// DockerLogs defines model for DockerLogs.
type DockerLogs struct {
	Logs []string `json:"logs"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
}

// HealthStatus defines model for HealthStatus.
type HealthStatus struct {
	Status HealthStatusStatus `json:"status"`

	// Version The agent version
	Version string `json:"version"`
}

// HealthStatusStatus defines model for HealthStatus.Status.
type HealthStatusStatus string

// Macro defines model for Macro.
type Macro struct {
	Command     string `json:"command"`
	Description string `json:"description"`
	Id          string `json:"id"`
	IsDangerous *bool  `json:"isDangerous,omitempty"`
	Name        string `json:"name"`
}

// MacroExecutionResult defines model for MacroExecutionResult.
type MacroExecutionResult struct {
	Output  string `json:"output"`
	Success bool   `json:"success"`
}

// MacroInput defines model for MacroInput.
type MacroInput struct {
	Command     string `json:"command"`
	Description string `json:"description"`
	IsDangerous *bool  `json:"isDangerous,omitempty"`
	Name        string `json:"name"`
}

// PowerActionRequest defines model for PowerActionRequest.
type PowerActionRequest struct {
	Action PowerActionRequestAction `json:"action"`
}

// PowerActionRequestAction defines model for PowerActionRequest.Action.
type PowerActionRequestAction string

// SystemLogEntry defines model for SystemLogEntry.
type SystemLogEntry struct {
	Id        string `json:"id"`
	Level     string `json:"level"`
	Message   string `json:"message"`
	Service   string `json:"service"`
	Timestamp string `json:"timestamp"`
}

// SystemMetrics defines model for SystemMetrics.
type SystemMetrics struct {
	Containers struct {
		Active  int  `json:"active"`
		Exited  int  `json:"exited"`
		Running bool `json:"running"`
		Warning int  `json:"warning"`
	} `json:"containers"`
	Cpu struct {
		Cores     int     `json:"cores"`
		Frequency float32 `json:"frequency"`
		Usage     float32 `json:"usage"`
	} `json:"cpu"`
	Disk struct {
		Free         int64   `json:"free"`
		Total        int64   `json:"total"`
		UsagePercent float32 `json:"usagePercent"`
		Used         int64   `json:"used"`
	} `json:"disk"`
	Memory struct {
		Free         int64   `json:"free"`
		Total        int64   `json:"total"`
		UsagePercent float32 `json:"usagePercent"`
		Used         int64   `json:"used"`
	} `json:"memory"`
	Network struct {
		DownMbps float32 `json:"downMbps"`
		UpMbps   float32 `json:"upMbps"`
	} `json:"network"`
	Power struct {
		Watts int `json:"watts"`
	} `json:"power"`
	Thermals *[]struct {
		Sensor      string  `json:"sensor"`
		Temperature float32 `json:"temperature"`
	} `json:"thermals,omitempty"`

	// Uptime Uptime in seconds
	Uptime int64 `json:"uptime"`
}

// ExecuteSystemCommandJSONBody defines parameters for ExecuteSystemCommand.
type ExecuteSystemCommandJSONBody struct {
	Command string `json:"command"`
}

// ExecutePowerActionJSONRequestBody defines body for ExecutePowerAction for application/json ContentType.
type ExecutePowerActionJSONRequestBody = PowerActionRequest

// ExecuteDockerContainerActionJSONRequestBody defines body for ExecuteDockerContainerAction for application/json ContentType.
type ExecuteDockerContainerActionJSONRequestBody = DockerActionRequest

// ExecuteSystemCommandJSONRequestBody defines body for ExecuteSystemCommand for application/json ContentType.
type ExecuteSystemCommandJSONRequestBody ExecuteSystemCommandJSONBody

// CreateSystemMacroJSONRequestBody defines body for CreateSystemMacro for application/json ContentType.
type CreateSystemMacroJSONRequestBody = MacroInput

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health check endpoint
	// (GET /health)
	GetHealth(ctx echo.Context) error
	// Execute a power action
	// (POST /power/action)
	ExecutePowerAction(ctx echo.Context) error
	// Get Docker containers
	// (GET /system/docker/containers)
	GetSystemDockerContainers(ctx echo.Context) error
	// Execute Docker container action
	// (POST /system/docker/containers/{id}/action)
	ExecuteDockerContainerAction(ctx echo.Context, id string) error
	// Get Docker container logs
	// (GET /system/docker/containers/{id}/logs)
	GetSystemDockerContainerLogs(ctx echo.Context, id string) error
	// Execute an arbitrary system command
	// (POST /system/execute)
	ExecuteSystemCommand(ctx echo.Context) error
	// Get system logs
	// (GET /system/logs)
	GetSystemLogs(ctx echo.Context) error
	// Get all macros
	// (GET /system/macros)
	GetSystemMacros(ctx echo.Context) error
	// Create a new macro
	// (POST /system/macros)
	CreateSystemMacro(ctx echo.Context) error
	// Delete a macro
	// (DELETE /system/macros/{id})
	DeleteSystemMacro(ctx echo.Context, id string) error
	// Execute a macro
	// (POST /system/macros/{id}/execute)
	ExecuteSystemMacro(ctx echo.Context, id string) error
	// Get real-time system metrics
	// (GET /system/metrics)
	GetSystemMetrics(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetHealth converts echo context to params.
func (w *ServerInterfaceWrapper) GetHealth(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetHealth(ctx)
	return err
}

// ExecutePowerAction converts echo context to params.
func (w *ServerInterfaceWrapper) ExecutePowerAction(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ExecutePowerAction(ctx)
	return err
}

// GetSystemDockerContainers converts echo context to params.
func (w *ServerInterfaceWrapper) GetSystemDockerContainers(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSystemDockerContainers(ctx)
	return err
}

// ExecuteDockerContainerAction converts echo context to params.
func (w *ServerInterfaceWrapper) ExecuteDockerContainerAction(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ExecuteDockerContainerAction(ctx, id)
	return err
}

// GetSystemDockerContainerLogs converts echo context to params.
func (w *ServerInterfaceWrapper) GetSystemDockerContainerLogs(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSystemDockerContainerLogs(ctx, id)
	return err
}

// ExecuteSystemCommand converts echo context to params.
func (w *ServerInterfaceWrapper) ExecuteSystemCommand(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ExecuteSystemCommand(ctx)
	return err
}

// GetSystemLogs converts echo context to params.
func (w *ServerInterfaceWrapper) GetSystemLogs(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSystemLogs(ctx)
	return err
}

// GetSystemMacros converts echo context to params.
func (w *ServerInterfaceWrapper) GetSystemMacros(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSystemMacros(ctx)
	return err
}

// CreateSystemMacro converts echo context to params.
func (w *ServerInterfaceWrapper) CreateSystemMacro(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateSystemMacro(ctx)
	return err
}

// DeleteSystemMacro converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteSystemMacro(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteSystemMacro(ctx, id)
	return err
}

// ExecuteSystemMacro converts echo context to params.
func (w *ServerInterfaceWrapper) ExecuteSystemMacro(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ExecuteSystemMacro(ctx, id)
	return err
}

// GetSystemMetrics converts echo context to params.
func (w *ServerInterfaceWrapper) GetSystemMetrics(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetSystemMetrics(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/health", wrapper.GetHealth)
	router.POST(baseURL+"/power/action", wrapper.ExecutePowerAction)
	router.GET(baseURL+"/system/docker/containers", wrapper.GetSystemDockerContainers)
	router.POST(baseURL+"/system/docker/containers/:id/action", wrapper.ExecuteDockerContainerAction)
	router.GET(baseURL+"/system/docker/containers/:id/logs", wrapper.GetSystemDockerContainerLogs)
	router.POST(baseURL+"/system/execute", wrapper.ExecuteSystemCommand)
	router.GET(baseURL+"/system/logs", wrapper.GetSystemLogs)
	router.GET(baseURL+"/system/macros", wrapper.GetSystemMacros)
	router.POST(baseURL+"/system/macros", wrapper.CreateSystemMacro)
	router.DELETE(baseURL+"/system/macros/:id", wrapper.DeleteSystemMacro)
	router.POST(baseURL+"/system/macros/:id/execute", wrapper.ExecuteSystemMacro)
	router.GET(baseURL+"/system/metrics", wrapper.GetSystemMetrics)

}